core: deprecate CoreAPI.Dht, introduce CoreAPI.Routing
