+++ client/rpc/api.go
@@ -227,6 +227,8 @@ func (api *HttpApi) Object() iface.ObjectAPI {
 	return (*ObjectAPI)(api)
 }
 
+// nolint deprecated
+// Deprecated: use [HttpApi.Routing] instead.
 func (api *HttpApi) Dht() iface.DhtAPI {
 	return (*DhtAPI)(api)
 }
