+++ core/coreiface/tests/routing.go
@@ -2,6 +2,7 @@ package tests
 
 import (
 	"context"
+	"io"
 	"testing"
 	"time"
 
@@ -23,6 +24,9 @@ func (tp *TestSuite) TestRouting(t *testing.T) {
 	t.Run("TestRoutingGet", tp.TestRoutingGet)
 	t.Run("TestRoutingPut", tp.TestRoutingPut)
 	t.Run("TestRoutingPutOffline", tp.TestRoutingPutOffline)
+	t.Run("TestRoutingFindPeer", tp.TestRoutingFindPeer)
+	t.Run("TestRoutingFindProviders", tp.TestRoutingFindProviders)
+	t.Run("TestRoutingProvide", tp.TestRoutingProvide)
 }
 
 func (tp *TestSuite) testRoutingPublishKey(t *testing.T, ctx context.Context, api iface.CoreAPI, opts ...options.NamePublishOption) (path.Path, ipns.Name) {
@@ -95,6 +99,148 @@ func (tp *TestSuite) TestRoutingPutOffline(t *testing.T) {
 	err = api.Routing().Put(ctx, ipns.NamespacePrefix+name.String(), data)
 	require.Error(t, err, "this operation should fail because we are offline")
 
-	err = api.Routing().Put(ctx, ipns.NamespacePrefix+name.String(), data, options.Put.AllowOffline(true))
+	err = api.Routing().Put(ctx, ipns.NamespacePrefix+name.String(), data, options.Routing.AllowOffline(true))
 	require.NoError(t, err)
+}
+
+func (tp *TestSuite) TestRoutingFindPeer(t *testing.T) {
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+	apis, err := tp.MakeAPISwarm(t, ctx, 5)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	self0, err := apis[0].Key().Self(ctx)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	laddrs0, err := apis[0].Swarm().LocalAddrs(ctx)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if len(laddrs0) != 1 {
+		t.Fatal("unexpected number of local addrs")
+	}
+
+	time.Sleep(3 * time.Second)
+
+	pi, err := apis[2].Routing().FindPeer(ctx, self0.ID())
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	if pi.Addrs[0].String() != laddrs0[0].String() {
+		t.Errorf("got unexpected address from FindPeer: %s", pi.Addrs[0].String())
+	}
+
+	self2, err := apis[2].Key().Self(ctx)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	pi, err = apis[1].Routing().FindPeer(ctx, self2.ID())
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	laddrs2, err := apis[2].Swarm().LocalAddrs(ctx)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if len(laddrs2) != 1 {
+		t.Fatal("unexpected number of local addrs")
+	}
+
+	if pi.Addrs[0].String() != laddrs2[0].String() {
+		t.Errorf("got unexpected address from FindPeer: %s", pi.Addrs[0].String())
+	}
+}
+
+func (tp *TestSuite) TestRoutingFindProviders(t *testing.T) {
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+	apis, err := tp.MakeAPISwarm(t, ctx, 5)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	p, err := addTestObject(ctx, apis[0])
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	time.Sleep(3 * time.Second)
+
+	out, err := apis[2].Routing().FindProviders(ctx, p, options.Routing.NumProviders(1))
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	provider := <-out
+
+	self0, err := apis[0].Key().Self(ctx)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	if provider.ID.String() != self0.ID().String() {
+		t.Errorf("got wrong provider: %s != %s", provider.ID.String(), self0.ID().String())
+	}
+}
+
+func (tp *TestSuite) TestRoutingProvide(t *testing.T) {
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+	apis, err := tp.MakeAPISwarm(t, ctx, 5)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	off0, err := apis[0].WithOptions(options.Api.Offline(true))
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	s, err := off0.Block().Put(ctx, &io.LimitedReader{R: rnd, N: 4092})
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	p := s.Path()
+
+	time.Sleep(3 * time.Second)
+
+	out, err := apis[2].Routing().FindProviders(ctx, p, options.Routing.NumProviders(1))
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	_, ok := <-out
+
+	if ok {
+		t.Fatal("did not expect to find any providers")
+	}
+
+	self0, err := apis[0].Key().Self(ctx)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	err = apis[0].Routing().Provide(ctx, p)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	out, err = apis[2].Routing().FindProviders(ctx, p, options.Routing.NumProviders(1))
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	provider := <-out
+
+	if provider.ID.String() != self0.ID().String() {
+		t.Errorf("got wrong provider: %s != %s", provider.ID.String(), self0.ID().String())
+	}
 }
