+++ docs/changelogs/v0.27.md
@@ -7,6 +7,8 @@
 - [Overview](#overview)
 - [🔦 Highlights](#-highlights)
   - [Gateway: support for `/api/v0` is deprecated](#gateway-support-for-apiv0-is-deprecated)
+  - [IPNS resolver cache's TTL can now be configured](#ipns-resolver-caches-ttl-can-now-be-configured)
+  - [RPC client: deprecated DHT API, added Routing API](#rpc-client-deprecated-dht-api-added-routing-api)
 - [📝 Changelog](#-changelog)
 - [👨‍👩‍👧‍👦 Contributors](#-contributors)
 
@@ -23,6 +25,12 @@ If you have a legacy software that relies on this behavior, and want to expose p
 #### IPNS resolver cache's TTL can now be configured
 
 You can now configure the upper-bound of a cached IPNS entry's Time-To-Live via [`Ipns.MaxCacheTTL`](https://github.com/ipfs/kubo/blob/master/docs/config.md#ipnsmaxcachettl).
+
+#### RPC client: deprecated DHT API, added Routing API
+
+The RPC client now includes a Routing API to match the available commands in `/api/v0/routing`. In addition, the DHT API has been marked as deprecated. 
+
+In the next version, all DHT deprecated methods will be removed from the Go RPC client.
 
 ### 📝 Changelog
