+++ docs/file-transfer.md
@@ -68,12 +68,12 @@ pitfalls that people run into)
 ### Checking providers
 When requesting content on ipfs, nodes search the DHT for 'provider records' to
 see who has what content. Let's manually do that on node B to make sure that
-node B is able to determine that node A has the data. Run `ipfs dht findprovs
+node B is able to determine that node A has the data. Run `ipfs routing findprovs
 <hash>`. We expect to see the peer ID of node A printed out. If this command
 returns nothing (or returns IDs that are not node A), then no record of A
 having the data exists on the network. This can happen if the data is added
 while node A does not have a daemon running. If this happens, you can run `ipfs
-dht provide <hash>` on node A to announce to the network that you have that
+routing provide <hash>` on node A to announce to the network that you have that
 hash. Then if you restart the `ipfs get` command, node B should now be able
 to tell that node A has the content it wants. If node A's peer ID showed up in
 the initial `findprovs` call, or manually providing the hash didn't resolve the
